Add Test File
Using Tape as the testing (micro?)framework.

Using JSDOM to create a DOM without a browser (as the test will be run 
using Node, not a browser)

Using the File System Module (fs) so the test can read the html file and 
therefore JSDOM can create a DOM.

Then put the DOM in the global scope, to make it globally available in 
the current node environment.

Finally, of-course, require the script file in.

I know it's against convention to add tests after writing the app... But 
I want to learn and move forward using TDD.